Add Switch Behavior to SlideToAct
Improved the logic behind the `animate_completion` attribute in order to
let the slider behave as a switch. Now the switch can be animated back
with a drag and the value of `isCompleted` is now consistent.

Moreover added the `complete_text` to provide a second @string to show
when the slider is complete. If provided, the slider will crossfade
between the two strings.

Fixes #72