The current version of SlideToAct cannot automatically adjust to the available space in the vertical direction.
This results in the LinearLayout, depending on the number of views displayed, or depending on the device resolution, not all views can be fully displayed, or free space remains unused.

This extension adds the missing dynamic height adjustment capability to SlideToAct.

Two changes are made here:
1. in "onMeasure()" the dynamic height adjustment capability is added.
2. in "onSizeChanged()" the condition "if (mBorderRadius == -1)" is removed.

Change 2. corrects the problem that when more free space becomes available, i.e. when SlideToAct becomes higher, the radius is not reapplied.
For Example: As a result, the slide button does not remain round, but becomes a square with rounded corners.